requirements.txt
asyncio
httpx
python-dotenv